+++ packages/component/src/block/accountStatus/sections/accountTotal.tsx
@@ -93,7 +93,7 @@ export const AccountTotal: FC<AccountTotalProps> = (props) => {
           </div>
           <Divider vertical className="px-3" />
 
-          <div className="border border-solid px-2 rounded border-primary-light text-primary-light text-sm h-[30px] leading-[30px]">
+          <div className="border border-solid px-2 rounded border-primary-light text-primary-light text-sm h-[30px] leading-[30px] flex items-center">
             {/* {`${new Decimal(currentLeverage).todp(2)}x`} */}
             <Numeral precision={2} surfix="x">
               {currentLeverage}