+++ packages/component/src/block/withdraw/sections/actionButton.tsx
@@ -67,7 +67,7 @@ export const ActionButton: FC<ActionButtonProps> = (props) => {
       return (
         <StatusGuardButton>
           <Button
-            className="min-w-[200px]"
+            fullWidth
             onClick={onWithdraw}
             disabled={props.disabled || props.loading}
             loading={loading}
@@ -81,7 +81,7 @@ export const ActionButton: FC<ActionButtonProps> = (props) => {
     if (chains?.length === 1) {
       return (
         <Button
-          className="min-w-[200px]"
+          fullWidth
           onClick={() => {
             const chain = chains[0];
             if (chain) {
@@ -101,7 +101,7 @@ export const ActionButton: FC<ActionButtonProps> = (props) => {
       );
     }
     return (
-      <Button className="min-w-[200px]" onClick={openChainPicker}>
+      <Button fullWidth onClick={openChainPicker}>
         Switch Network
       </Button>
     );
@@ -124,7 +124,7 @@ export const ActionButton: FC<ActionButtonProps> = (props) => {
       )}
 
       <div className="flex justify-center">
-        <div className="py-3">{actionButton}</div>
+        <div className="py-3 min-w-[200px]">{actionButton}</div>
       </div>
     </>
   );