+++ packages/component/src/stories/hooks/useAccount.stories.tsx
@@ -7,7 +7,7 @@ import { MemoryConfigStore } from "@orderly.network/core";
 import { API } from "@orderly.network/types";
 
 const AccountDemo: FC<{
-  info: API.AccountInfo;
+  info: any;
 }> = (props) => {
   return (
     <div className="text-black">
@@ -32,6 +32,6 @@ export default meta;
 export const Default: Story = {
   render: () => {
     const account = useAccount();
-    return <AccountDemo info={account.info} />;
+    return <AccountDemo info={account.state} />;
   },
 };