+++ packages/component/src/tab/tabContext.tsx
@@ -1,3 +1,4 @@
+import { useLocalStorage } from "@orderly.network/hooks";
 import React, { FC, PropsWithChildren, useCallback, useState } from "react";
 
 export interface TabContextState {
@@ -9,8 +10,14 @@ export interface TabContextState {
 
 const TabContext = React.createContext<TabContextState>({} as TabContextState);
 
-const TabContextProvider: FC<PropsWithChildren<{ data?: any }>> = (props) => {
-  const [visible, setVisible] = useState(true);
+const TabContextProvider: FC<
+  PropsWithChildren<{
+    data?: any;
+    collapsed: boolean;
+    onToggleCollapsed?: () => void;
+  }>
+> = (props) => {
+  // const [visible, setVisible] = useState<boolean>(() => props.collapsed);
   const [data, setData] = useState<any>(props.data || {});
 
   const updateData = useCallback((key: string, value: any) => {
@@ -25,9 +32,10 @@ const TabContextProvider: FC<PropsWithChildren<{ data?: any }>> = (props) => {
   return (
     <TabContext.Provider
       value={{
-        contentVisible: visible,
+        contentVisible: !props.collapsed,
         toggleContentVisible: () => {
-          setVisible((visible) => !visible);
+          // setVisible((visible: boolean) => !visible);
+          props.onToggleCollapsed?.();
         },
         data,
         updateData,