+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # storybook
 
+## 1.0.40
+
+### Patch Changes
+
+- update
+
+## 1.0.39
+
+### Patch Changes
+
+- update
+
 ## 1.0.38
 
 ### Patch Changes