+++ packages/component/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @orderly.network/components
 
+## 0.0.45
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.45
+  - @orderly.network/utils@0.0.39
+  - @orderly.network/net@1.0.43
+
+## 0.0.44
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.44
+  - @orderly.network/utils@0.0.38
+  - @orderly.network/net@1.0.42
+
 ## 0.0.43
 
 ### Patch Changes