+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orderly.network/hooks
 
+## 0.0.46
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/futures@0.0.46
+  - @orderly.network/types@0.0.45
+  - @orderly.network/utils@0.0.39
+  - @orderly.network/core@0.0.45
+  - @orderly.network/net@1.0.43
+
+## 0.0.45
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/futures@0.0.45
+  - @orderly.network/types@0.0.44
+  - @orderly.network/utils@0.0.38
+  - @orderly.network/core@0.0.44
+  - @orderly.network/net@1.0.42
+
 ## 0.0.44
 
 ### Patch Changes