+++ packages/net/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @orderly.network/net
 
+## 1.0.43
+
+### Patch Changes
+
+- update
+
+## 1.0.42
+
+### Patch Changes
+
+- update
+
 ## 1.0.41
 
 ### Patch Changes