+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @orderly.network/utils
 
+## 0.0.39
+
+### Patch Changes
+
+- update
+
+## 0.0.38
+
+### Patch Changes
+
+- update
+
 ## 0.0.37
 
 ### Patch Changes